改变了main

+++ src/main/java/com/tthome/visneymanager/entity/Size.java
@@ -7,6 +7,6 @@ import java.io.Serializable;
 @Data
 public class Size implements Serializable {
     private static final long serialVersionUID = 4141327978368950732L;
-    private Short sizeId;
+    private Integer sizeId;
     private String siziMsg;
 }
